+++ src/azure-cli/HISTORY.rst
@@ -14,7 +14,10 @@ Release History
 **Compute**
 
 * vm create: Add --enable-agent configuration.
+* vm create: Use standard public IP SKU automatically when using zones.
+* vm create: Compose a valid computer name from VM name if computer name is not provided.
 * vmss create: Add --computer-name-prefix parameter to support custom computer name prefix of virtual machines in the VMSS.
+* Update galleries API version to 2019-07-01.
 
 **Network**
 
@@ -67,6 +70,7 @@ Release History
 **IoT**
 
 * Add new routing source type: DigitalTwinChangeEvents
+* Fix #2116: Unexpected 'az iot hub show' error for resource not found.
 
 **Key Vault**